structure dtos and private methods
